requirements.txt
apache-beam==2.48.0
apache-flink==1.18.0
apache-flink-libraries==1.18.0
avro-python3==1.10.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
fastavro==1.9.7
fasteners==0.19
find_libpython==0.4.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
kafka-python==2.0.2
numpy==1.24.4
objsize==0.6.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pemja==0.3.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zstandard==0.23.0
